move loading of maintenance events into data service
