+++ src/app/maintenance-overview/maintenance-overview.component.ts
@@ -15,32 +15,11 @@ import { formatQueryDate } from '../model/base';
   templateUrl: './maintenance-overview.component.html',
   styleUrl: './maintenance-overview.component.css'
 })
-export class MaintenanceOverviewComponent implements OnInit {
-  protected maintenanceEvents?: Array<IncidentResponseData>
+export class MaintenanceOverviewComponent {
+
   constructor(
     private incidentService: IncidentService,
     protected config: AppConfigService,
-    protected dataService: DataService
+    protected data: DataService
   ) {}
-
-  async ngOnInit(): Promise<void> {
-    const currentDate = dayjs();
-    const future = currentDate.add(this.config.maintenancePreviewDays, "d");
-
-    const incidents = (await firstValueFrom(this.incidentService.getIncidents(
-      formatQueryDate(currentDate),
-      formatQueryDate(future)
-    )))?.data;
-
-    this.maintenanceEvents = incidents.filter((incident) => {
-      incident.affects = incident.affects?.filter((affects) => {
-        const maintenanceSeverity = this.config.severities.get('maintenance');
-        const maintenanceSeverityValue = maintenanceSeverity ? maintenanceSeverity.end : 0;
-
-        return affects.severity !== undefined && affects.severity <= maintenanceSeverityValue;
-      });
-
-      return incident.affects !== undefined && incident.affects.length > 0;
-    });
-  }
 }
